requirements.txt
Flask==3.1.1
python-keycloak==5.5.1
WTForms==3.2.1
